+++ src/ApiClient.php
@@ -34,6 +34,7 @@ final class ApiClient
 {
     use Endpoints\Taxonomies;
     use Endpoints\Auth;
+    use Endpoints\Products;
 
     public const VERSION = '0.6.0';